feat: use AWS Go SDK v2
