+++ internal/app/assume.go
@@ -4,12 +4,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
+	"strings"
 
 	"github.com/aripalo/vegas-credentials/internal/assumecfg"
 	"github.com/aripalo/vegas-credentials/internal/config/locations"
 	"github.com/aripalo/vegas-credentials/internal/credentials"
 	"github.com/aripalo/vegas-credentials/internal/msg"
-	"github.com/aripalo/vegas-credentials/internal/totp"
+	"github.com/aripalo/vegas-credentials/internal/prompt"
 	"github.com/dustin/go-humanize"
 )
 
@@ -31,19 +33,29 @@ func (a *App) Assume(flags AssumeFlags) error {
 
 	creds := credentials.New(cfg)
 
-	if err = creds.FetchFromCache(); err != nil {
+	if err = creds.Load(); err != nil {
 		msg.Debug("ℹ️", fmt.Sprintf("Credentials: Cached: %s", err))
 		msg.Debug("ℹ️", "Credentials: STS: Fetching...")
 		msg.Debug("ℹ️", fmt.Sprintf("MFA: TOTP: %s", cfg.MfaSerial))
 
 		// TODO refactor this
-		t := totp.New(totp.TotpOptions{
-			YubikeySerial: cfg.YubikeySerial,
-			YubikeyLabel:  cfg.YubikeyLabel,
-			EnableGui:     !a.NoGui,
-		})
+		/*
+					t := totp.New(totp.TotpOptions{
+						YubikeySerial: cfg.YubikeySerial,
+						YubikeyLabel:  cfg.YubikeyLabel,
+						EnableGui:     !a.NoGui,
+					})
 
-		err = creds.FetchFromAWS(creds.BuildProvider(t.Get))
+			    creds.BuildProvider(t.Get)
+		*/
+
+		fmt.Fprintln(os.Stderr, "CODE:")
+		code, err := prompt.Cli(context.Background(), "")
+		if err != nil {
+			panic(err) // TODO fix
+		}
+
+		err = creds.New(strings.TrimSpace(code))
 		if err != nil {
 			if errors.Is(err, context.DeadlineExceeded) {
 				msg.Fatal(fmt.Sprintf("Operation Timeout"))
