--- internal/credentials/model.go
@@ -1,24 +1,0 @@
-package credentials
-
-import (
-	"io"
-	"time"
-
-	"github.com/aripalo/vegas-credentials/internal/assumecfg"
-	"github.com/aripalo/vegas-credentials/internal/cache"
-	"github.com/aripalo/vegas-credentials/internal/credentials/sts"
-)
-
-// Credentials defines the output format expected by AWS credential_process
-// https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-sourcing-external.html
-type Credentials struct {
-	cfg             assumecfg.AssumeCfg
-	provider        sts.Provider
-	repo            cache.Repository
-	output          io.Writer
-	Version         int       `json:"Version"`
-	AccessKeyID     string    `json:"AccessKeyId"`
-	SecretAccessKey string    `json:"SecretAccessKey"`
-	SessionToken    string    `json:"SessionToken"`
-	Expiration      time.Time `json:"Expiration"`
-}
